Separate main CSS module into header and content ones
