+++ app/components/views/TransactionPage/TransactionPage.module.css
@@ -6,275 +6,7 @@
   padding: 0;
   top: 157px;
 }
-.indicatorMixin,
-.indicatorPending,
-.indicatorConfirmed {
-  display: inline-block;
-  padding: 5px 8px 5px 20px;
-  border-style: solid;
-  border-width: 1px;
-  border-radius: 3px;
-  font-size: 12px;
-  line-height: 8px;
-  text-align: right;
-  text-transform: capitalize;
-  background-position: 6px 50%;
-  background-size: 10px;
-  background-repeat: no-repeat;
-}
-.indicatorPending {
-  border-color: var(--input-color);
-  background-image: var(--indicator-pending-icon);
-  color: var(--input-color);
-}
-.indicatorConfirmed {
-  border-color: var(--accent-color);
-  background-image: var(--indicator-confirmed-icon);
-  color: var(--accent-color);
-}
-.details {
-  padding-left: 80px;
-  font-size: 13px;
-  margin-bottom: 50px;
-}
-.top {
-  padding-left: 80px;
-  padding-top: 25px;
-  padding-bottom: 28px;
-  font-size: 13px;
-  background-color: var(--txdetails-top-bg);
-}
-.io {
-  padding-left: 80px;
-  padding-top: 28px;
-}
-.topRow {
-  margin-bottom: 9px;
-  display: flex;
-}
-.name {
-  color: var(--tx-detail-text);
-  margin-right: 11px;
-  min-width: 90px;
-  text-align: right;
-  font-size: 13px;
-  line-height: 17px;
-}
-.value {
-  word-wrap: break-word;
-  display: flex;
-}
-.value > a {
-  color: var(--input-color-default);
-  text-shadow: rgba(0, 0, 0, 0.01) 0 0 1px;
-}
-.value.nonFlex {
-  display: block;
-}
-.valueRawTx {
-  font-size: 13px;
-  line-height: 15px;
-  width: 540px;
-  height: 233px;
-  overflow-y: auto;
-  background-color: var(--background-back-color);
-  padding: 10px;
-  color: var(--stroke-color-hovered);
-}
-.overview {
-  margin-bottom: 3em;
-  display: flex;
-  justify-content: space-between;
-  width: 659px;
-}
-.overviewTitleMixin,
-.overviewTitleCreated,
-.overviewTitleConsumed,
-.overviewTitleEmpty {
-  color: var(--input-color-default);
-  font-size: 19px;
-  line-height: 24px;
-  margin-bottom: 13px;
-}
-.overviewTitleEmpty {
-  color: var(--disabled-color);
-}
-.inputArrow {
-  background-image: var(--tx-details-arrow-left);
-  background-position: 100% 50%;
-  background-size: 37px;
-  background-repeat: no-repeat;
-  float: right;
-  height: 43px;
-  width: 43px;
-  margin-top: 3em;
-}
-.inputs,
-.outputs {
-  width: 220px;
-}
-.inputArea,
-.outputArea {
-  width: 180px;
-  background-color: var(--background-back-color);
-  padding: 15px 20px 1px 20px;
-  margin-bottom: 10px;
-}
-.outputArea {
-  margin-right: 180px;
-}
-.direction {
-  display: block;
-  width: 100%;
-  height: 28px;
-  padding-top: 4px;
-  padding-left: 80px;
-  float: left;
-  clear: left;
-  font-size: 20px;
-  line-height: 19px;
-  text-align: left;
-}
-.row {
-  margin-bottom: 15px;
-}
-.subtitle {
-  display: flex;
-  margin-top: 22px;
-  flex-wrap: wrap;
-}
-.subtitlePair {
-  display: flex;
-  margin-top: 3px;
-}
-.subtitleSentfrom {
-  margin-right: 5px;
-}
-.subtitleDate,
-.subtitleAccount,
-.address.nonWallet,
-.address {
-  font-size: 13px;
-  line-height: 17px;
-  padding: 2px 10px 2px 10px;
-  border-radius: 3px;
-  background-color: var(--blue-highlight-background);
-  color: var(--input-color-default);
-  text-overflow: ellipsis;
-  overflow: hidden;
-  white-space: nowrap;
-  text-align: center;
-  margin-right: 5px;
-}
-.subtitleDate {
-  max-height: 22px;
-}
-.subtitleAccount {
-  width: 80px;
-  max-height: 22px;
-}
-.address.nonWallet {
-  width: 155px;
-  margin-bottom: 5px;
-}
-.address {
-  width: 80px;
-  margin-bottom: 5px;
-}
-.amount {
-  font-size: 17px;
-  line-height: 27px;
-  width: 100%;
-}
-.valueText {
-  font-weight: 400;
-}
-.title {
-  color: var(--stroke-color-hovered);
-  font-size: 20px;
-  line-height: 25px;
-  margin-bottom: 7px;
-}
-.last {
-  border-bottom: 0px none transparent;
-}
-@media screen and (max-width: 1106px) {
-  .overview {
-    width: 633px;
-  }
-  .io,
-  .details,
-  .top {
-    padding-left: 36px;
-  }
-  .value-rawtx {
-    width: 511px;
-  }
-}
-@media screen and (max-width: 1106px) {
-  .overview {
-    width: 633px;
-  }
-  .io,
-  .details,
-  .top {
-    padding-left: 36px;
-  }
-  .value-rawtx {
-    width: 511px;
-  }
-}
 @media screen and (max-width: 768px) {
-  .overview {
-    flex-direction: column;
-    padding-left: 42px;
-  }
-  .details,
-  .top {
-    padding-left: 30px;
-  }
-  .inputArrow {
-    transform: rotate(90deg);
-    margin: 20px 0 25px 88px;
-  }
-  .top .value > a {
-    text-overflow: ellipsis;
-    width: 220px;
-    white-space: nowrap;
-    overflow: hidden;
-  }
-  .details .topRow {
-    width: 335px;
-  }
-  .details .topRow .value > a {
-    word-wrap: break-word;
-    display: inline-block;
-    width: 223px;
-  }
-  .details .topRow.rowTransaction {
-    flex-direction: column;
-  }
-  .details .topRow.rowTransaction .name {
-    text-align: left;
-  }
-  .details .topRow.rowTransaction .value {
-    margin-top: 10px;
-    margin-left: -20px;
-    width: 354px;
-    position: relative;
-  }
-  .details .topRow.rowTransaction .value .valueRawTx {
-    width: 335px;
-  }
-  .details
-    .topRow.rowTransaction
-    .value
-    .clipboardBox.receiveContentNestCopyToClipboardIcon {
-    position: absolute;
-    top: -28px;
-    right: 5px;
-    margin: 0;
-  }
   .standalonePage > div:first-child {
     padding-bottom: 10px;
   }
@@ -282,16 +14,10 @@
     display: block;
     margin-left: 30px;
   }
-  .standalonePage .subtitle {
-    margin-top: 10px;
-  }
   .standalonePage > div:last-child {
     padding: 0 0 40px 0;
     position: relative;
     top: initial;
     overflow-y: initial;
-  }
-  .standalonePage .subtitle {
-    width: 355px;
   }
 }
